feature: user menu
- mobile view fixes

+++ scripts/deploy.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
+DEPLOY_SERVER=$1
+if [ -z $DEPLOY_SERVER ];
+then
+  echo No server provided
+  exit
+fi
+
 mkdir -p publish
 rm -rf publish/*
 # build
-dotnet build -c Release -r linux-arm64 backend/HomeSpace.Api/HomeSpace.Api.csproj || exit
+dotnet build -c Release -r linux-arm64 --self-contained backend/HomeSpace.Api/HomeSpace.Api.csproj || exit
 # publish
-dotnet publish -c Release -r linux-arm64 backend/HomeSpace.Api/HomeSpace.Api.csproj -o publish/
+dotnet publish -c Release -r linux-arm64 --no-build backend/HomeSpace.Api/HomeSpace.Api.csproj -o publish/
 
 #go to client folder
 cd client || exit 
@@ -17,7 +24,6 @@ cd .. || exit
 cp -r client/dist publish/wwwroot
 cp backend/home-space.service publish/
 
-DEPLOY_SERVER=$1
 # copy published to rpi
 # scp publish/* ubuntu@192.168.1.144:./home-space-cs/
 rsync -av --progress --delete --exclude=appsettings.Production.json  publish/  ubuntu@"$DEPLOY_SERVER":./home-space/
